+++ packages/toolbars-new/web/src/modals/heading/UpdateHeadingPanel.tsx
@@ -34,13 +34,17 @@ const UpdateHeadingPanel = ({
   return (
     <div className={styles.panel}>
       <div
-        className={classNames(styles.panel_row, styles.buttonComponent)}
-        onClick={() => onApply(false)}
-        onKeyDown={e => onKeyDown(e, onApply, onClose, true)}
+        className={styles.panel_row}
         // eslint-disable-next-line jsx-a11y/no-noninteractive-tabindex
         tabIndex={0}
       >
-        {t('FormattingToolbar_TextStyle_Heading_Apply', { optionName })}
+        <div
+          className={classNames(styles.panel_row, styles.buttonComponent)}
+          onClick={() => onApply(false)}
+          onKeyDown={e => onKeyDown(e, onApply, onClose, true)}
+        >
+          {t('FormattingToolbar_TextStyle_Heading_Apply', { optionName })}
+        </div>
       </div>
       <div className={styles.separator} />
       <div
@@ -67,11 +71,7 @@ const UpdateHeadingPanel = ({
         tabIndex={resetButtonDisabled ? -1 : 0}
       >
         <div
-          className={classNames(
-            styles.panel_row,
-            styles.buttonComponent,
-            resetButtonDisabled && styles.disabled
-          )}
+          className={classNames(styles.buttonComponent, resetButtonDisabled && styles.disabled)}
           onClick={() => !resetButtonDisabled && onReset()}
           onKeyDown={e => !resetButtonDisabled && onKeyDown(e, onReset, onClose)}
         >